[New] Complex filters
The filteredPokemon array is populated by filtering the pokemonMapped array based on the selected filters.

The pokemonToHide array is populated by filtering the pokemonMapped array with Pokémon that are not present in the filteredPokemon array.

The Pokémon cards with IDs matching the numbers of Pokémon in the pokemonToHide array are given the "hidden" class to hide them.